% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ppos_2arm_bin.R
\name{calculate_ppos_2arm_bin}
\alias{calculate_ppos_2arm_bin}
\title{Function to calculate the posterior probability of success for a two-arm trial with binary endpoints based on Bayesian go no-go dual-criterion design}
\usage{
calculate_ppos_2arm_bin(
  n1,
  n2,
  x1,
  x2,
  alpha_trt_prior,
  beta_trt_prior,
  alpha_ctrl_prior,
  beta_ctrl_prior,
  N1,
  N2,
  cv1,
  cv2,
  sig_level
)
}
\arguments{
\item{n1}{Number of patients in treatment arm}

\item{n2}{Number of patients in control arm}

\item{x1}{Number of responses in treatment arm}

\item{x2}{Number of responses in control arm}

\item{alpha_trt_prior}{shape parameter 1 of Beta prior for treatment arm}

\item{beta_trt_prior}{shape parameter 2 of Beta prior for treatment arm}

\item{alpha_ctrl_prior}{shape parameter 1 of Beta prior for control arm}

\item{beta_ctrl_prior}{shape parameter 2 of Beta prior for control arm}

\item{N1}{Total number of patients in treatment arm}

\item{N2}{Total number of patients in control arm}

\item{cv1}{Benchmark for the first criterion}

\item{cv2}{Benchmark for the second criterion}

\item{sig_level}{Significance level for the first of dual-criterion}
}
\value{
Posterior probability of success, defined as the probability that the dual-criterion is met at final given the interim data.
}
\description{
Function to calculate the posterior probability of success for a two-arm trial with binary endpoints based on Bayesian go no-go dual-criterion design
}
\examples{
x1 = 16; x2 = 9; n1 = 25;
          n2 = 25; N1 = 32; N2 = 32;
          alpha_trt_prior = 1; beta_trt_prior = 1;
          alpha_ctrl_prior = 1; beta_ctrl_prior = 1;
          cv1 = 0.0; cv2 = 0.4; sig_level = 0.9
          calculate_ppos_2arm_bin(
            n1, n2, x1, x2, alpha_trt_prior, beta_trt_prior,
            alpha_ctrl_prior, beta_ctrl_prior, N1, N2, cv1, cv2,
            sig_level
          )
}
