% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ppos_single_arm_bin.R
\name{calculate_ppos_single_arm_bin}
\alias{calculate_ppos_single_arm_bin}
\title{Function to calculate the posterior probability of success for a single-arm trial with binary endpoint
based on Bayesian go no-go dual-criterion design}
\usage{
calculate_ppos_single_arm_bin(
  x = 13,
  n = 16,
  alpha_prior = 1,
  beta_prior = 1,
  N = 26,
  cv1 = 0.4,
  cv2 = 0.6,
  sig_level = 0.95,
  n_sims = 10000
)
}
\arguments{
\item{x}{interim number of responses}

\item{n}{interim sample size}

\item{alpha_prior}{shape parameter 1 of Beta prior}

\item{beta_prior}{shape parameter 2 of Beta prior}

\item{N}{final sample size}

\item{cv1}{benchmark for  criterion 1}

\item{cv2}{benchmark for  criterion 2}

\item{sig_level}{significance level for criterion 1}

\item{n_sims}{iterations for Monte Carlo simulation}
}
\value{
Posterior probability of success, defined as the probability that the dual-criterion is met at final given the interim data.
}
\description{
Function to calculate the posterior probability of success for a single-arm trial with binary endpoint
based on Bayesian go no-go dual-criterion design
}
\examples{
x <- 13
n <- 16
alpha_prior <- 1
beta_prior <- 1
N <- 26
cv1 <- 0.4
cv2 <- 0.6
sig_level <- 0.95
n_sims <- 10000
seed <- 42
ppos <- calculate_ppos_single_arm_bin(x, n, alpha_prior, beta_prior, N, cv1, cv2, sig_level, n_sims)
print(ppos)
}
