% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2arm_trial_decision.R
\name{make_2arm_trial_decision}
\alias{make_2arm_trial_decision}
\title{Title: Make 2-arm trial decision for Binary Endpoints}
\usage{
make_2arm_trial_decision(
  n1,
  n2,
  x1,
  x2,
  alpha_trt_prior,
  beta_trt_prior,
  alpha_ctrl_prior,
  beta_ctrl_prior,
  N1,
  N2,
  cv1,
  cv2,
  sig_level = 0.9,
  ppos_go = 0.9,
  ppos_no_go = 0.1,
  orr_trt = 0.4,
  orr_ctl = 0.2,
  n_final_iterations = 1000
)
}
\arguments{
\item{n1}{interim sample size for treatment arm}

\item{n2}{interim sample size for control arm}

\item{x1}{observed number of events in treatment arm}

\item{x2}{observed number of events in control arm}

\item{alpha_trt_prior}{shape parameter 1 for beta prior for treatment arm}

\item{beta_trt_prior}{shape parameter 2 for beta prior for treatment arm}

\item{alpha_ctrl_prior}{shape parameter 1 for beta prior for control arm}

\item{beta_ctrl_prior}{shape parameter 2 for beta prior for control arm}

\item{N1}{final sample size for treatment arm}

\item{N2}{final sample size for control arm}

\item{cv1}{delta_orr benchmark for criterion 1}

\item{cv2}{delta_orr benchmark for criterion 2}

\item{sig_level}{significance level for dual-criterion}

\item{ppos_go}{early go cutoff for PPoS}

\item{ppos_no_go}{early no go cutoff for PPoS}

\item{orr_trt}{assumed ORR for treatment arm}

\item{orr_ctl}{assumed ORR for control arm}

\item{n_final_iterations}{number of samples to draw from the posterior distribution}
}
\value{
a data frame with interim decision, final decision, and total sample size
}
\description{
Title: Make 2-arm trial decision for Binary Endpoints
}
\examples{
n1 = 25; n2 = 25; x1 = 10; x2 = 5;
          alpha_trt_prior = 1; beta_trt_prior = 1;
          alpha_ctrl_prior = 1; beta_ctrl_prior = 1;
          N1 = 32; N2 = 32; cv1 = 0.0; cv2 = 0.23;
          sig_level = 0.90; ppos_go = 0.90; ppos_no_go = 0.10;
          orr_trt = 0.4; orr_ctl = 0.2;
          n_final_iterations = 10000
          make_2arm_trial_decision(n1, n2, x1, x2,
                        alpha_trt_prior, beta_trt_prior,
                        alpha_ctrl_prior, beta_ctrl_prior,
                        N1, N2, cv1, cv2, sig_level,
                        ppos_go, ppos_no_go, orr_trt, orr_ctl,
                        n_final_iterations)
}
