% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_OC_2arm_binary.R
\name{simulate_OC_2arm_binary}
\alias{simulate_OC_2arm_binary}
\title{Title: Simulate 2-arm trial with binary endpoints per PPoS and GNG dual-criterion design}
\usage{
simulate_OC_2arm_binary(
  n_simulation = 1000,
  n1 = 25,
  n2 = 25,
  N1 = 32,
  N2 = 32,
  orr_trt = 0.4,
  orr_ctl = 0.2,
  ppos_go = 0.9,
  ppos_no_go = 0.1,
  sig_level = 0.9,
  cv1 = 0,
  cv2 = 0.23,
  alpha_trt_prior = 1,
  beta_trt_prior = 1,
  alpha_ctrl_prior = 1,
  beta_ctrl_prior = 1,
  seed = 112024
)
}
\arguments{
\item{n_simulation}{number of simulations}

\item{n1}{interim sample size for treatment arm}

\item{n2}{interim sample size for control arm}

\item{N1}{final sample size for treatment arm}

\item{N2}{final sample size for control arm}

\item{orr_trt}{assumed ORR for treatment arm}

\item{orr_ctl}{assumed ORR for control arm}

\item{ppos_go}{cutoff for PPoS to early go at interim}

\item{ppos_no_go}{cutoff for PPoS to early no-go at interim}

\item{sig_level}{significance level for  criterion 1}

\item{cv1}{benchmark for response difference criterion 1}

\item{cv2}{benchmark for response difference criterion 2}

\item{alpha_trt_prior}{shape parameter 1 for beta prior for treatment arm}

\item{beta_trt_prior}{shape parameter 2 for beta prior for treatment arm}

\item{alpha_ctrl_prior}{shape parameter 1 for beta prior for control arm}

\item{beta_ctrl_prior}{shape parameter 2 for beta prior for control arm}

\item{seed}{seed for reproducibility}
}
\value{
a data frame with interim decision, final decision, and expected sample size
}
\description{
Title: Simulate 2-arm trial with binary endpoints per PPoS and GNG dual-criterion design
}
\examples{
n_simulation = 10; n1 = 25; n2 = 25;
          N1 = 32; N2 = 32; orr_trt = 0.4; orr_ctl = 0.2;
          ppos_go = 0.90; ppos_no_go = 0.10; sig_level = 0.90;
          cv1 = 0.0; cv2 = 0.23; alpha_trt_prior = 1; beta_trt_prior = 1;
          alpha_ctrl_prior = 1; beta_ctrl_prior = 1;
          seed = 112024
          simulate_OC_2arm_binary(n_simulation, n1, n2, N1, N2, orr_trt, orr_ctl,
                       ppos_go, ppos_no_go, sig_level, cv1, cv2,
                       alpha_trt_prior, beta_trt_prior, alpha_ctrl_prior, beta_ctrl_prior,
                       seed)
}
